requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
fastapi==0.115.2
filelock @ file:///C:/Pavan/MyCodingWorld/doc.ai_llama32-11B_fastapi_docker/llama32-11B-Instruct-service/install_files/filelock-3.16.1-py3-none-any.whl#sha256=2082e5703d51fbf98ea75855d9d5527e33d8ff23099bec374a134febee6946b0
frozenlist==1.4.1
fsspec @ file:///C:/Pavan/MyCodingWorld/doc.ai_llama32-11B_fastapi_docker/llama32-11B-Instruct-service/install_files/fsspec-2024.10.0-py3-none-any.whl#sha256=03b9a6785766a4de40368b88906366755e2819e758b83705c88cd7cb5fe81871
greenlet==3.1.1
grpcio==1.67.0
grpcio-tools==1.67.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
hyperframe==6.0.1
idna==3.10
Jinja2 @ file:///C:/Pavan/MyCodingWorld/doc.ai_llama32-11B_fastapi_docker/llama32-11B-Instruct-service/install_files/jinja2-3.1.4-py3-none-any.whl#sha256=bc5dd2abb727a5319567b7a813e6a2e7318c39f4f487cfe6c89c6f9c7d25197d
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-qdrant==0.1.4
langchain-text-splitters==0.3.0
langsmith==0.1.136
MarkupSafe @ file:///C:/Pavan/MyCodingWorld/doc.ai_llama32-11B_fastapi_docker/llama32-11B-Instruct-service/install_files/MarkupSafe-3.0.2-cp312-cp312-win_amd64.whl#sha256=8e06879fc22a25ca47312fbe7c8264eb0b662f6db27cb2d3bbbc74b1df4b9b87
marshmallow==3.23.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx @ file:///C:/Pavan/MyCodingWorld/doc.ai_llama32-11B_fastapi_docker/llama32-11B-Instruct-service/install_files/networkx-3.4.2-py3-none-any.whl#sha256=df5d4365b724cf81b8c6a7312509d0c22386097011ad1abe274afd5e9d3bbc5f
numpy @ file:///C:/Pavan/MyCodingWorld/doc.ai_llama32-11B_fastapi_docker/llama32-11B-Instruct-service/install_files/numpy-2.1.2-cp312-cp312-win_amd64.whl#sha256=456e3b11cb79ac9946c822a56346ec80275eaf2950314b249b512896c0d2505e
orjson==3.10.9
packaging==24.1
pillow @ file:///C:/Pavan/MyCodingWorld/doc.ai_llama32-11B_fastapi_docker/llama32-11B-Instruct-service/install_files/pillow-11.0.0-cp312-cp312-win_amd64.whl#sha256=8ec4a89295cd6cd4d1058a5e6aec6bf51e0eaaf9714774e1bfac7cfc9051db47
portalocker==2.10.1
propcache==0.2.0
protobuf==5.28.2
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pypdf==5.0.1
python-dotenv==1.0.1
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.12
pywin32==308
PyYAML==6.0.2
qdrant-client==1.12.0
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
setuptools==75.2.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.1
torch @ file:///C:/Pavan/MyCodingWorld/doc.ai_llama32-11B_fastapi_docker/llama32-11B-Instruct-service/install_files/torch-2.5.0-cp312-cp312-win_amd64.whl#sha256=ea718746469246cc63b3353afd75698a288344adb55e29b7f814a5d3c0a7c78d
torchaudio @ file:///C:/Pavan/MyCodingWorld/doc.ai_llama32-11B_fastapi_docker/llama32-11B-Instruct-service/install_files/torchaudio-2.5.0-cp312-cp312-win_amd64.whl#sha256=e98ad102e0e574a397759078bc9809afc05de6e6f8ac0cb26d2245e0d3817adf
torchvision @ file:///C:/Pavan/MyCodingWorld/doc.ai_llama32-11B_fastapi_docker/llama32-11B-Instruct-service/install_files/torchvision-0.20.0-cp312-cp312-win_amd64.whl#sha256=bb0da0950d2034a0412c251a3a9117ff9612157f45177d37ba1b20b472c0864b
tqdm==4.66.5
transformers==4.45.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wheel==0.44.0
yarl @ file:///C:/Pavan/MyCodingWorld/doc.ai_llama32-11B_fastapi_docker/llama32-11B-Instruct-service/install_files/yarl-1.16.0-cp312-cp312-win_amd64.whl#sha256=1208ca14eed2fda324042adf8d6c0adf4a31522fa95e0929027cd487875f0240
